Move download components into download directory
